requirements/core.txt
matplotlib>=3.3.0
numpy>=1.20.2
pytorch>=1.8.0
tensorboard>=2.4.1
torchsummary>=1.5.1
torchvision>=0.9.0
